Add basic Taskwarrior JSON parsing
Add `Taskwarrior.Task` struct definition and related builder function.

Since the JSON data is assumed to be originating from Taskwarrior's
`task export`, the code assumes general correctness of the data, meaning
that it is rather optimistic.

Due to this optimism, change the implementation for
`Taskwarrior.from_json/1` to expect the returned term to be a list of
elements instead of an `{:ok, tasks}` tuple.