Add support for UDAs
The approach in this patch expects the user to supply the list UDAs,
instead of the UDAs being extracted automatically.

One could probably go through each exported task and filter for keys
that are not part of the Taskwarrior "columns", as seen in the output of
`task columns` (note: this command will include UDA columns if one has
them configured). This automatic approach may be better, but for the
time being I'll run with explicit UDA extraction.